// libs/@innomobile/mapbox/src/lib/class/mapbox.interface.ts
export interface MapboxDirectionsOptions {
    api?: string;
    profile?: string;
    
    alternatives?: boolean;
    annotations?: string;
    approaches?: string;
    banner_instructions?: boolean;
    geometries?:  string;
    language?: string;
    overview?:  string;
    steps?: boolean;

    access_token?: string;
}